Windows x64 warning fix and a query-assembly hardening. UI_Records_Line was passing its size_t through Com_sprintf's int size parameter (C4267 on 64-bit MSVC; GCC does not flag this conversion) -- explicit cast, matching the codebase convention. And the sortable-column lookup now returns the whitelist entry itself: DB_Top and DB_TopFormat assemble their SQL from the entry's own field and table strings, so no caller-supplied byte reaches the SQL text even on a successful whitelist match. The player-facing rank command was already safe (full-string match required before any assembly); this closes the gap between safe and provably-inert.

+++ ui_boards.c
@@ -142,7 +142,7 @@ static qboolean UI_Records_Line(char *out, size_t outsize,
 	if (!rec->holder[0])
 		return false;
 
-	Com_sprintf(out, outsize, "%s: %s - %d", label, rec->holder, rec->value);
+	Com_sprintf(out, (int)outsize, "%s: %s - %d", label, rec->holder, rec->value);
 	return true;
 }
